requirements.txt
openpyxl
Pillow
cv2
pytesseract
